+++ include/rangelib/rangelib.hpp
@@ -1,12 +1,12 @@
 /* /////////////////////////////////////////////////////////////////////////
- * File:        rangelib/rangelib.hpp
+ * File:    rangelib/rangelib.hpp
  *
- * Purpose:     Root header for the STLSoft implementation of RangeLib.
+ * Purpose: Root header for the STLSoft implementation of RangeLib.
  *
- * Created:     12th October 2004
- * Updated:     11th March 2024
+ * Created: 12th October 2004
+ * Updated: 9th July 2024
  *
- * Home:        http://stlsoft.org/
+ * Home:    http://stlsoft.org/
  *
  * Copyright (c) 2019-2024, Matthew Wilson and Synesis Information Systems
  * Copyright (c) 2004-2019, Matthew Wilson and Synesis Software
@@ -48,7 +48,7 @@
 # define RANGELIB_VER_RANGELIB_HPP_RANGELIB_MAJOR    1
 # define RANGELIB_VER_RANGELIB_HPP_RANGELIB_MINOR    4
 # define RANGELIB_VER_RANGELIB_HPP_RANGELIB_REVISION 4
-# define RANGELIB_VER_RANGELIB_HPP_RANGELIB_EDIT     40
+# define RANGELIB_VER_RANGELIB_HPP_RANGELIB_EDIT     41
 #endif /* !STLSOFT_DOCUMENTATION_SKIP_SECTION */
 
 /** \file rangelib/rangelib.hpp
@@ -64,7 +64,7 @@
  * components.
  *
  * The major version is denoted by the _RANGELIB_VER_MAJOR preprocessor symbol.
- * A changes to the major version component implies that a dramatic change has
+ * A change to the major version component implies that a dramatic change has
  * occurred in the libraries, such that considerable changes to source dependent
  * on previous versions would need to be effected.
  *
@@ -123,12 +123,13 @@
 # define _RANGELIB_VER_1_1_1    0x00010101  /*!< Version 1.1.1 (with STLSoft 1.9.1) */
 # define _RANGELIB_VER_1_1_2    0x00010102  /*!< Version 1.1.2 (with STLSoft 1.9.25) */
 # define _RANGELIB_VER_1_1_3    0x010103ff  /*!< Version 1.1.3 (with STLSoft 1.9.113) */
+# define _RANGELIB_VER_1_1_4    0x010103ff  /*!< Version 1.1.4 (with STLSoft 1.11.1 alpha 8) */
 #endif /* !STLSOFT_DOCUMENTATION_SKIP_SECTION */
 
 #define _RANGELIB_VER_MAJOR     1
 #define _RANGELIB_VER_MINOR     1
-#define _RANGELIB_VER_REVISION  3
-#define _RANGELIB_VER           _RANGELIB_VER_1_1_3
+#define _RANGELIB_VER_REVISION  4
+#define _RANGELIB_VER           _RANGELIB_VER_1_1_4
 
 
 /* /////////////////////////////////////////////////////////////////////////